distsqlrun: move sorterstrategy to sorter
No code changes except for code movement.

Release note: None
